feature: Added an additional button to toolbar to copy an invitation link to clipboard
The invitation link is built with the the scheme: 'https://' + infinityContext.infinityHost + '/webapp/#/?conference=mp' + infinityContext.conferenceAlias + '&pin=' + infinityContext.conferencePin.
Currently the link will allow a join as host. This should be changed in future.

The interface InfinityContext was introduced for PIN, Alias and Host information

The copy-to-clipboard package is used: https://www.npmjs.com/package/copy-to-clipboard

Refs: GI-60