fix: During on hold state the video of all participants will be muted now
The muteVideo option by uuid was added by R&D to the @pexip/infinity 16.1.0" package. It is used now to mute the video of all participants during hold state. An upgrade to version 16.1.0 was necessary. The new version brings small API extensions but no API breaks. It was necessary to adjust the infinityClientMock in toolbar/Toolbar.test.tsx.

Refs: GI-50 | GI-76