feature: Added a mute video icon to toolbar
The agent can now mute the video via toolbar. The icon gets red if the camera is muted.
The icon also gets red if the camera is muted by the hold state.
The self view will be hidden if the video is muted by toolbar icon.
Currently a reference of the self view will be provided to the toolbar.
I think we should introduce a global state management to decouple the components and to allow to react on server events based on a single source of truth

Refs: GI-71