fix: (genesysService)  Sometimes on hold fails
There seems to be an edge case where Genesys produces an events with two participants with the purpose "agent". One has the state "terminated" and one has the state "connected".
We just filter for a participant with the purpose "agent" and when we choose the one with the state "terminated" the on hold flag will never be set. An do not trigger a process in our application.
The logic filters for active agents now.

Refs: GI-112