fix(app): The call can transferred from one agent to another and vice versa in same interaction context
Added a listener logic for a connect event.
This event is triggered if a call is accepted by the agent.
If the app state is NO_ACTIVE_CALL and a connect event was detected the app will rejoin the conference.
To archive this pexipAgentPin and pexipNode are defined as global now

Refs: GI-145